v2, commit log: link events in a blockchain-style way
