+++ core/src/crypto/utils.rs
@@ -1,5 +1,4 @@
 use crate::models::Base64EncodedText;
-use base64::engine::fast_portable::NO_PAD;
 use rand::{distributions::Alphanumeric, Rng};
 use sha2::{Digest, Sha256};
 use uuid::Uuid;
