+++ core/src/node/mod.rs
@@ -1,3 +1,2 @@
 pub mod db;
 pub mod server_api;
-pub mod app;
