+++ meta-server-emulator/src/lib.rs
@@ -1,1 +1,1 @@
-mod server;
+pub mod server;
