+++ meta-server-emulator/src/server/mod.rs
@@ -1,2 +1,2 @@
-mod kv_store;
-mod meta_server;
+pub mod kv_store;
+pub mod meta_server;
