+++ meta-test/Cargo.toml
@@ -6,3 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+meta-secret-core = { path = "../core" }
+meta-server-emulator = { path = "../meta-server-emulator" }
+
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
